Negative votes - bug fix
The vote function doesn't handle negative votes correctly, in particular, the piece that checks "can the voter add a vote to this proposal," is flawed.

If x = number of votes allocated to a proposal, the function checks if x^2 - (x + 1)^2 <= number of credits remaining. So if you're adding negative votes, x + 1 is going in the wrong direction.